backend_deploy/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
google-generativeai==0.8.3
python-dotenv==1.0.1
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator
a2wsgi==1.10.0
